[refactor] TheModals를 common components 에서 module의 modal-frame으로 변경
