+++ tsconfig.app.json
@@ -10,6 +10,7 @@
       "@common/*": ["./src/common/*"],
       "@core/*": ["./src/core/*"],
       "@^header/*": ["./src/modules/header/*"],
+      "@^modal-frame/*": ["./src/modules/modal-frame/*"],
       "@^login/*": ["./src/modules/login/*"],
       "@^stuffs/*": ["./src/modules/stuffs/*"],
       "@^histories/*": ["./src/modules/histories/*"],
